Rename orchestration names constant
